+++ packages/api/src/promise/Combinator.ts
@@ -2,31 +2,27 @@
 // This software may be modified and distributed under the terms
 // of the Apache-2.0 license. See the LICENSE file for details.
 
+import { UnsubFunction } from './types';
+
 import { isFunction } from '@polkadot/util';
 
 export type CombinatorCallback = (value: Array<any>) => any;
 export type CombinatorFunction = {
-  (cb: (value: any) => any): Promise<any>,
-  unsubscribe?: (subscriptionsId: number) => Promise<any>
+  (cb: (value: any) => any): UnsubFunction | any
 };
-
-let combinatorId = 5000;
-const allCombinators: { [index: number]: Combinator } = {};
 
 export default class Combinator {
   protected _allHasFired: boolean = false;
   protected _callback: CombinatorCallback;
   protected _fired: Array<boolean> = [];
   protected _fns: Array<CombinatorFunction> = [];
-  protected _id: number = ++combinatorId;
+  protected _isActive: boolean = true;
   protected _results: Array<any> = [];
-  protected _subscriptionIds: Array<Promise<number>> = [];
+  protected _unsubscriptions: Array<UnsubFunction | any> = [];
 
   constructor (fns: Array<CombinatorFunction | [Array<any>, CombinatorFunction]>, callback: CombinatorCallback) {
-    allCombinators[this._id] = this;
-
     this._callback = callback;
-    this._subscriptionIds = fns.map((input, index): Promise<number> => {
+    this._unsubscriptions = fns.map((input, index) => {
       const [args, fn] = Array.isArray(input)
         ? input
         : [[], input];
@@ -37,22 +33,6 @@ export default class Combinator {
       // @ts-ignore Not quite 100% how to have a variable number at the front here
       return fn(...args, this.createCallback(index));
     });
-  }
-
-  static lookup (id: number): Combinator {
-    return allCombinators[id];
-  }
-
-  static unsubscribe (id: number): Promise<any> {
-    const combinator = Combinator.lookup(id);
-
-    return combinator
-      ? combinator.unsubscribe()
-      : Promise.resolve(false);
-  }
-
-  get id (): number {
-    return this._id;
   }
 
   protected allHasFired (): boolean {
@@ -73,7 +53,7 @@ export default class Combinator {
   }
 
   protected triggerUpdate (): void {
-    if (!isFunction(this._callback) || !this.allHasFired()) {
+    if (!this._isActive || !isFunction(this._callback) || !this.allHasFired()) {
       return;
     }
 
@@ -84,22 +64,13 @@ export default class Combinator {
     }
   }
 
-  unsubscribe (): Promise<any> {
-    delete allCombinators[this._id];
+  unsubscribe (): void {
+    this._isActive = false;
 
-    return Promise.all(
-      this._subscriptionIds.map((subscriptionPromise, index) => {
-        const unsubscribe = this._fns[index].unsubscribe;
-
-        return !unsubscribe
-          ? Promise.resolve(true)
-          : subscriptionPromise
-            .then((subscriptionId) =>
-              unsubscribe(subscriptionId)
-            )
-            .then(() => true)
-            .catch(() => false);
-      })
-    );
+    this._unsubscriptions.forEach((unsubscribe) => {
+      if (isFunction(unsubscribe)) {
+        unsubscribe();
+      }
+    });
   }
 }