+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+# 0.36.1
+
+Api Promise has been updated in the way we deal with subscriptions. Previously a subscription returned `Promise<number>` where the caller was to keep track of the id and use it in subsequent unsubscribes. Now any subscriptions return an unsubscribe/destroy function `(): void`, that is use for removing the subscriptions, e.g.
+
+```js
+const unsubscribe = api.query.balance.freeBalanceOf(Alice, (balance) => {...});
+
+unsubscribe(); // here we destroy the subscription
+```
+
 # 0.35.1
 
 Swapped to new metadata structures from Substrate. If the API is not working with your node, update Substrate to latest master branch. (Or 0.9.1 for Charred Cherry). Dropped support for old metadata as found as far back as BBQ Birch.